refactor: extract podcast episode player tests from podcast detail page tests
